initialize a quic repo
As mentioned in https://github.com/lxin/tls_hs#the-backgrounds: "some
people may argue that TLS handshake should stay in user space and use
up-call to user space in kernel to complete the handshake". The repo
is to implement the idea.

- The userspace handshake part for the in-kernel QUIC is in lib/.
- The kernel part for the rest of QUIC protocol is in net/quic/.

Note that the main part of the QUIC protocol is still in Kernel space.

Signed-off-by: Xin Long <lucien.xin@gmail.com>