simplify the socket context by QUIC_SOCKOPT_CONTEXT sockopt
When setup the socket context, it's better to have all parameters in
one whole sockopt to make sure the setup not half succeed, as well
as make programming clear and simply. struct quic_context containing
all the context params is added for this purpose.

While at it, also add a sample app to use in-kernel QUIC socket with
with pre-defined keys, instead of getting from ngtcp2 handshake. And
This app can run without ngtcp2 built.

Signed-off-by: Xin Long <lucien.xin@gmail.com>