protocol: add proto ops for handshake
Now we get many checks for handshake packets with quic_handshake_user()
in quic_sendmsg(), quic_recvmsg() and quic_do_rcv(), as handshake and
common packets basically go to different processing paths.

This patch is to separate the processing by introducing struct proto for
handshake, quicv6_handshake_prot, where it calls corresponding functions
during the handshake.

Note that when the socket is created, the handshake proto is set to the
sock by default. Once the context is set by setsockopt, it changes to
the common proto for common packets processing.

Signed-off-by: Xin Long <lucien.xin@gmail.com>