socket: do not send ack frame after setting context for client
On client side, it indicates the handshake is done when connection_done
frame is received in userspace. However, in kernel space it has no idea
the packet number of the connection_done packet.

It is not correct to ACK it with number = 0 after setting context, as it
may ACK another packet by accident while in fact it doesn't receive that
packet in kernel space.

Note that the rule is all Short packets must be received from the kernel
space. If any is received in userspace during handshake, then do not ACK
it in userspace. Later the retransmitted one will arrive in the kernel
space.

Signed-off-by: Xin Long <lucien.xin@gmail.com>