frame: add support for token management
This patch is to add QUIC_SOCKOPT_NEW_TOKEN to allow users sending
QUIC_FRAME_NEW_TOKEN frame to the peer.

When receiving a frame of QUIC_SOCKOPT_NEW_TOKEN, currently it only
saves the token to sock so that later the user can get it by sockopt
QUIC_SOCKOPT_TOKEN for the future handshake.

Signed-off-by: Xin Long <lucien.xin@gmail.com>