frame: add basic processing for unsupported frames
With return 0 in the unsupported frame processing, it continues parsing
the next frame without skipping all this frame in skb->data and may get
an unexpected frame type.

This patch returns -EOPNOTSUPP in processing these frames that it does
not support and can not skip, and skip all frame data in these that it
does not support but can skip.

Signed-off-by: Xin Long <lucien.xin@gmail.com>