timer: Add an idle timer
Section 10.1 of the QUIC RFC describes the Idle Timeout timer. If a
max_idle_timeout is specified by either endpoint in its transport
parameters, the connection is silently closed and its state is discarded
when it remains idle for longer than the minimum of the max_idle_timeout
values advertised by both endpoints.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>