stream: add support for stream management
This patch enhanced the stream state transition, and add two socket
options:

  QUIC_SOCKOPT_STREAM_RESET
  QUIC_SOCKOPT_STREAM_STOP_SENDING

to allow users to stop peer sending msg or reset local stream.

Meanwhile, the creating and processing for four types of frames are
added:

  QUIC_FRAME_MAX_STREAMS_UNI
  QUIC_FRAME_MAX_STREAMS_BIDI
  QUIC_FRAME_RESET_STREAM
  QUIC_FRAME_STOP_SENDING

This will complete the stream management.

While at it, fix the stream_id type to u64 and enrich the stream_flags
in sndinfo.

Signed-off-by: Xin Long <lucien.xin@gmail.com>