tests: add ipv6 tests in perf_test and sample_test
Instead of adding -4/6 in argv, it uses getaddrinfo() to get the
address family.

Note that it failed in IPv6 testing from msquic to lkquic, as in
gnutls it doesn't accept the server_name ext with ipv6 address
while msquic uses ipv6 address as server_name by default.

But it won't be a problem in real user case, as it always uses
the server's hostname instead ipv6 address.

Signed-off-by: Xin Long <lucien.xin@gmail.com>