crypto: support all four ciphers in kernel
All four Ciphers (AES-128-GCM, AES-256-GCM, AES-128-CCM, CHACHA20-POLY1305)
are added in kernel here. struct quic_cipher is introduced to adapt secret
and key length as well as the aead, skc and shash name for them with a
special process for ccm in payload_en/decrypt and a special process for
chacha20 in header_en/decrypt.

Some variables from linux/tls.h are reused to set up the secret context,
and also QUIC_KEY_LEN and QUIC_SECRET_LEN are adjusted to be able to hold
the secret and key for all kinds of ciphers.

Note that struct quic_crypto_secret changes due to the QUIC_SECRET_LEN's
change, which would cause uapi's change.

Signed-off-by: Xin Long <lucien.xin@gmail.com>