packet: use correct max_udp_payload_size to calculate mss
The maximum UDP payload size parameter is an integer value that limits
the size of UDP payloads that the endpoint is willing to receive. So
it should use the one in remote transport param, which is set in outq
max_udp_payload_size.

While at it, pass both local and remote params to quic_cong_get_param
and quic_cong_set_param, in case of the future use.

Signed-off-by: Xin Long <lucien.xin@gmail.com>