packet: fix the packet number type to s64
As it says in rfc9000#name-packet-numbers, the packet number is an integer
in the range 0 to 2^62-1. If the packet number for sending reaches 2^62-1,
the sender MUST close the connection without sending a CONNECTION_CLOSE
frame or any further packets.

So change the packet number type from u32 to s64.

Currently, it keeps the number length as 4 in the packet header, when the
number > U32_MAX, as it says in rfc9000#name-packet-number-encoding-and-,
the number of bits required to represent the packet number is reduced by
including only the least significant bits of the packet number.

Note this patch was tested by setting QUIC_PN_MAP_BASE_PN to different
values on both server and client sides.

Signed-off-by: Xin Long <lucien.xin@gmail.com>