handshake: refactor the userspace handshake
Since most of QUIC implementation is already in kernel, we can extract
a small part of ngtcp2 into libquic and drop the dependence of ngtcp2
library.

It not only reduces one library dependence for user to build, but also
brings us more control on the short packets handling in userspace and
works around the handshake failure issue caused by sending 1rtt too
early in the latest version of ngtcp2.

Note that we also changed some APIs name to make them more generic.

Signed-off-by: Xin Long <lucien.xin@gmail.com>