handshake: make the priority per connection
All connections share the global priority now, but it might change in
get_priority for each connection, and cause the handshake to use the
cipher from another connection.

So fix it by making the priority per connection. Note that as session
will be freed, we need to dup the peercerts data in x509_verify() to
return to users.

While at it, also fix a session leak by calling gnutls_deinit() in
quic_conn_destroy().

Signed-off-by: Xin Long <lucien.xin@gmail.com>