protocol: move long packets build and handle into kernel
This is a big change in both user handshake and kernel protocol.
It simplifies user handshake a lot and makes it only create and
process TLS messages. The LONG packets will be parsed and
created in kernel space.

In this way all QUIC related code will stay in kernel space, which
keeps the QUIC implementation as a whole, and avoids the layering
violation as Nick noticed.

Signed-off-by: Xin Long <lucien.xin@gmail.com>