packet: add retry packet and token support
Address verification is supported by adding retry packet processing
in this patch where retry token is created and carried in the retry
packet to reply to the initial packet from client and client will
send a new initial packet with this token for verification on the
server side.

The regular token is another type of token created and carried in
NEW_TOKEN frame sent from server to client so that next time the
client can carry it directly in the initial packet for verification.

Signed-off-by: Xin Long <lucien.xin@gmail.com>