crypto: use a global random_data to generate token and ticket_key
This patch creates a global random_data, token and ticket_key will
be generated with this random_data and client address. In this way
we can avoid sending the local random_data directly to client as
token, and avoid having the same ticket_key for different clients.

More importantly, when the QUIC server application crash/restart,
the token or session will still work.

Signed-off-by: Xin Long <lucien.xin@gmail.com>