packet: add stateless reset support
stateless_reset token is generated by quic_crypto_generate_token()
with "stateless_reset" label. stateless_reset is added to struct
quic_transport_param to enable it.

When receiving a non-initial packet on listen socket, it means the
connection for this packet might be gone, a stateless_reset packet
will be sent back to peer to reset the connection(we just close it
for now), so currently it can only be sent from server side. Also,
quic_unhash() is removed so that client socket can be found by the
stateless_reset packet with no connection id in it.

Note as the packet triggered stateless_reset is likely to be a
short packet, the socket lookup must search the socket hashtable
if no socket is found in the connection id hashtable, so that the
listen socket can respond to it with a stateless_reset packet.

Signed-off-by: Xin Long <lucien.xin@gmail.com>