packet: fix a null pointer dereference in quic_packet_transmit
In quic_packet_transmit(), when head->len + skb->len >= mss, the head
skb will be sent out and set to NULL. However, instead of setting the
current skb to head, it continues adding skb to head's frag_list,
which will cause a crash.

This patch is to fix it by setting it to packet->head and jumping to
try sending out the head skb. While at it, fix ipfragok to not allow
app packets to do fragmentation in quic_packet_config().

Signed-off-by: Xin Long <lucien.xin@gmail.com>