packet: start the ack timer to send ACKs to handshake packets
Currently, after receiving a handshake packet, it relies on the response
handshake packets from user space to bundle and flush this ACK.

However, the user space sends the response handshake packets only after
receiving sufficient data which could be included in multiple packets.
The peer in some implementation may send the next handshake packet
only after receiving the ACK to the last one.

So in case that no handshake packets from user space come to bundle it,
this patch start the ack timer to flush it out. Note the timer will be
stopped if any handshake packet is sent out.

Signed-off-by: Xin Long <lucien.xin@gmail.com>