socket: move udp_sk and port into struct quic_path_src
The members in src and dst of socket are not really different, and more
different members will be added in the future, like these for plpmtud
in dst, which shouldn't be in src.

This patch define quic_path_src and quic_path_dst wrapping quic_path_addr
for src and dst according, and also move udp_sk and port from socket to
quic_path_src.

Signed-off-by: Xin Long <lucien.xin@gmail.com>