crypto: sync payload encryption by calling crypto_wait_req
We currently don't support async de/encryption for payload either,
and crypto_wait_req() has to be called to wait for de/encryption
to complete. crypto->async_wait is added as the completion.

Note that we don't make it sync by using CRYPTO_ALG_ASYNC in mask
argument crypto_alloc_aead(), as it somehow impairs performance
quite a lot.

Signed-off-by: Xin Long <lucien.xin@gmail.com>