timer: do not send packets in timers if the socket is closed
When a socket is closing, it stops all timers but does not wait for
the already running timer. So in the timer handler, the socket may
be closed already, in which case it must not send out any packets.
Otherwise, a crash might be caused:

  [] BUG: kernel NULL pointer dereference, address: 0000000000000010
  [] RIP: 0010:quic_v4_lower_xmit+0x4c/0x150 [quic]
  [] Call Trace:
  []  <IRQ>
  []  quic_packet_transmit+0x10b/0x270 [quic]
  []  quic_packet_flush+0x22/0x80 [quic]
  []  quic_timer_delay_ack_timeout+0x85/0x150 [quic]
  []  call_timer_fn+0x26/0x130
  []  run_timer_softirq+0x1fa/0x460

Fix it by adding a check to goto out in some timer handlers if the
socket is closed.

Signed-off-by: Xin Long <lucien.xin@gmail.com>