uapi: delete the annotations in quic.h
The annotations in quic.h are more like user manual doc, we will create
a mandoc for it in the next patch, so delete all annotations in quic.h
in this patch.

While at it, move up the QUIC_SOCKOPT_ALPN and QUIC_SOCKOPT_SESSION_TICKET
definition, as it is not only used by handshake.

Signed-off-by: Xin Long <lucien.xin@gmail.com>