handshake: add IPPROTO_QUIC and SOL_QUIC in netinet/quic.h
Normally SOL_QUIC and IPPROTO_QUIC are defined in netinet/in.h, which is
from glibc. If the kernel is supported quic, but glibc is not updated,
users still can't find SOL_QUIC and IPPROTO_QUIC.

To work around this problem, it adds IPPROTO_QUIC and SOL_QUIC definition
into netinet/quic.h if they're not found, like netinet/sctp.h.

Signed-off-by: Xin Long <lucien.xin@gmail.com>