path: set path_alt for packets in rcv path
When replying to a packet, its src and dst addrs should use the dst
and src addrs from the original packet, instead of the active ones.

This patch adds path_alt in quic_rcv_cb to save the alt addr info,
so that the reply will also use the alt path if the original packet
was using it.

While at it, use QUIC_PATH_ALT_SRC/DST to replace the magic numbers
and extract the path validation code into quic_outq_validate_path().

Signed-off-by: Xin Long <lucien.xin@gmail.com>