output: reset data_blocked flag in ack to the block frame
Currently the data_blocked flag is reset by MAX_DATA while the
DATA_BLOCKED frame is ACKed by ACK frame. If the MAX_DATA is
received but the ACK is lost, the DATA_BLOCKED frame is still
in RTX queue, but as the data_blocked flag is reset and more
DATA_BLOCKED frames will be sent and enqueued into RTX queue.

Multiple DATA_BLOCKED frames in RTX queue are not what we want,
so simplify it by moving the data_blocked flag reset in the ACK
processing.

Signed-off-by: Xin Long <lucien.xin@gmail.com>