crypto: support sending key update and crypto errcodes
This patch adds quic_outq_transmit_close() to send errcodes in
close frame to peer, and it is called in quic_packet_process()
and quic_packet_handshake_process(). errcode and frame are
collected in pki when processing the packet and passed into
quic_outq_transmit_close().

We only support QUIC_TRANSPORT_ERROR/KEY_UPDATE_CRYPTO in this
patch. While at it, fix the name of enum quic_transport_error.

Signed-off-by: Xin Long <lucien.xin@gmail.com>