socket: change the checks order in quic_wait_for_packet
Currently, when timeo is 0 it always returns -EAGAIN when there is
no skb in receive queue. However, when the socket is closed or
there is sk_err set, we expect to return the corresponding err.

To achieve that, we need to move the check for sk_state and sk_err
ahead of timeo in quic_wait_for_packet().

Signed-off-by: Xin Long <lucien.xin@gmail.com>