handshake: allow no certificate validation on client
Allow no certificate validation on client so that it will be easier
for users to create their certificate for testing.

With this patch, any certificate (like self-signed) will pass the
validation during handshake  when certificate_request is set to 3
on client.

While at it, delete "Transport Error Codes for All Error Cases" from
TODO in README, as it's been added in previous patches.

Signed-off-by: Xin Long <lucien.xin@gmail.com>