socket: add support for compatible versions negotiation
This patch adds support for compatible versions negotiation, which
is defined in rfc9368.

As there's only two existing versions, this is more like to add a
framework for now. The static quic_versions[][] includes all the
current supported versions and their compatible versions.

Note disable_compatible_version is introduced to transport param
to allow users to disable this feature and use the negotiation
defined in rfc8999 only.

Signed-off-by: Xin Long <lucien.xin@gmail.com>