crypto: carry original dcid in the token with aead encrypted
When validate_peer_address is set, instead of creating and enqueuing
a request sock, the server sends a retry packet with a cookie/token
to respond client's request. Then the client send a new request with
this token.

However, currently there's no info saved in the cookie/token for the
original dcid. As the server doesn't save any info locally for the
1st client request either, it cannot include original dcid from the
1st client request into the transport param server's initial packet.

To fix this, the patch saves the original dcid into the token of the
retry packet, and encrypts it with a local generated aead key in
quic_crypto_generate_token(), and decrypts and gets it with the same
aead key when the token is back with the 2nd client initial packet
in quic_crypto_verify_token().

Some unit tests are also added for these two functions.

Signed-off-by: Xin Long <lucien.xin@gmail.com>