frame: support preferred_address transport param on client side
When preferred_address transport param is received, it gets the
addr, conn_id and token, then it sets addr to the alt path and
adds conn_id and token to dest connection id set. pref_addr in
outqueue is also set to indicate preferred_address is ready.

When HANDSHAKE_DONE frame is received, quic_sock_change_daddr()
is called with addr == NULL to perform the preferred_address
change if pref_addr in outqueue is set.

In quic_sock_change_daddr(), if addr == NULL it skips the alt
address update and directly does a connection migration with
the alt peer address set previously.

Signed-off-by: Xin Long <lucien.xin@gmail.com>