socket: support reuseport for listen sockets
reuseport APIs are called in quic_hash() and quic_sock_lookup() on
listen sock to support reuseport.

With this feature, more than one processes can listen on the sockets
with the same address and port after enabling it via SO_REUSEPORT
socket option, the listen socket will be chosen according to the
value of hash(daddr) to process the incoming packet.

Note that currently QUIC socket supports 'reuseaddr' by default.
There's no need to call SO_REUSEADDR socket option if you want
to bind to the same address and port on multiple sockets.

Signed-off-by: Xin Long <lucien.xin@gmail.com>