input: add support for v4 mapped v6 address
This patch supports v4 mapped v6 address or ipv6 socket processing
ipv4 packets.

In quic_accept(), req->sa.v4.sin_family should be used to create the
accept socket, as we expect to create v4 socket to accept the ipv4
connection when it is a v6 listen socket.

In quic_do_listen_rcv(), the skb af_ops is set to for listen sk when
processing the packets, as we expect to use ipv4 APIs to send retry
packet for an ipv4 connection even if it is an ipv6 listen socket.

With this patch, users can use ipv6 socket binding to :: to accept
both ipv6 and ipv4 connections.

Signed-off-by: Xin Long <lucien.xin@gmail.com>