Makefile: add Makefile for quic module build
libquic is being moved to ktls-utils repo, a Makefile for
quic module build will be needed. Add it now so that the
reviewers from ktls-utils can use it at the moment.

With this patch, the quic.ko can be built directly by:

  # make

and installed by:

  # make install

Before libquic is removed from this repo, ./configure will
still rewrite the Makefile, and users need to use old cmds
from README to build and install quic module

Signed-off-by: Xin Long <lucien.xin@gmail.com>