sample_test: add support for psk test
psk module param is added to set the psk key_serial_t when modprobe
quic_sample_test.

For client, key_serial_t must be an available key from keyring and
it will be passed from to kernel to userspace for choose the psk.

For server, key_serial_t doesn't need to be a specific key (as long
as it's a non-zero value to enter psk mode). When processing the
request from client, it does key search from userspace, and no key
is needed to pass from kernel.

Signed-off-by: Xin Long <lucien.xin@gmail.com>