frame: validate dgram frame len on send and recv paths
As it says in RFC9221:

  DATAGRAM frames cannot be fragmented.

Therefore, it must fail to create frame in quic_frame_datagram_create()
and return -EINVAL in quic_sendmsg().

And also,

  An endpoint that receives a DATAGRAM frame that is larger than the
  value it sent in its max_datagram_frame_size transport parameter MUST
  terminate the connection with an error of type PROTOCOL_VIOLATION.

Thus, set frame->errcode to QUIC_TRANSPORT_ERROR_PROTOCOL_VIOLATION in
quic_frame_datagram_process() if that occurs.

Note that mss needs to be updated by quic_packet_mss_update() when it's
processing max_datagram_frame_size.

While at it, return -1 in quic_packet_config() when it returns err in
quic_packet_route(), and check the return value of quic_packet_config()
in all places.

Signed-off-by: Xin Long <lucien.xin@gmail.com>