cong: make congestion control comply more with the rfc
Use the similar names with these in rfc9002#section-appendix.b for the
members in struct quic_cong_ops for better understanding.

These is no chance from cong_avoid or recovery directly to slow_start
state, so delete the related code and unused parameters.

In quic_reno_on_packet_lost(), drop the packet loss check code as it's
already done before calling this function.

Add recovery_time to struct quic_cong_ops to save the time in entering
the recovery state, and use it to check if the packet acked are sent
in recovery state.

While at it, update the related tests in quic_unit_test and replace
transmit_ts with sent_time in struct quic_frame.

Signed-off-by: Xin Long <lucien.xin@gmail.com>