libquic: sync the libquic code from ktls-utils
As it does for QUIC handshake in ktls-utils, this patch is to split
the libquic.c to client.c server.c and handshake.c to make the code
more readable. Also, keep most code consistent with the handshake
code for QUIC in ktls-utils to make maintenance easy.

Note that there is no APIs change for users in this patch.

Signed-off-by: Xin Long <lucien.xin@gmail.com>