socket: set sk_state to closed in quic_disconnect
quic_disconnect() is only called when shutdown a listen socket
in QUIC. Set sk_state to closed in quic_disconnect so that a
blocking accept can be canceled when shutdown a listen sock.

In this case, quic_is_listen() will not work in quic_unhash(),
so change to check sk_max_ack_backlog instead.

Signed-off-by: Xin Long <lucien.xin@gmail.com>