socket: allow to automatically resize the sender buffer
With this patch, for applications that have not explicitly set the socket
send buffer size, the kernel will allow the send buffer to grow with the
congestion window (cwnd) and utilize the available bandwidth of the link.
If the network memory load on the server increases, the kernel will also
reduce the senders' windows.

Note that due to no advertised window carried in ACK frame, the Dynamic
right-sizing (DRS) can not be applied to QUIC, which means the receiver
buffer or window is always fixed (a "big enough" window).

While at it, initialize sk_snd/rcvbuf with sysctl_quic_r/wmem and expose
them in /proc/net/quic, as well as fix some indents related to conn_id.

Signed-off-by: Xin Long <lucien.xin@gmail.com>